requirements.txt
# FastAPI & Uvicorn
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Store & AI
langchain==0.1.4
langchain-community==0.0.19
pgvector==0.2.4
sentence-transformers==2.3.1

# HTTP Client
httpx==0.27.0
aiohttp==3.9.1

# Cache
redis==5.0.1
hiredis==2.3.2

# Authentication
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# Utils
python-dotenv==1.0.0
